add constants, code refactoring
